requirements.txt
appdirs==1.4.3
asgiref==3.3.0
boto3==1.16.22
botocore==1.19.22
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.9
django-allauth==0.44.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-instagram==0.3.2
django-mathfilters==1.0.0
django-storages==1.10.1
djangorestframework==3.12.2
gunicorn==20.0.4
html5lib==1.0.1
humanize==2.6.0
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
lxml==4.6.1
msgpack==0.6.2
numpy==1.19.4
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.0.0
progress==1.5
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-google-places==1.4.2
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
